requirements.txt
Package          Version   
---------------- ----------
altgraph         0.17      
auto-py-to-exe   2.6.6     
bottle           0.12.18   
bottle-websocket 0.2.9     
cffi             1.14.0    
Click            7.0       
Eel              0.11.0    
future           0.18.2    
gevent           1.4.0     
gevent-websocket 0.10.1    
greenlet         0.4.15    
pefile           2019.4.18 
Pillow           7.1.1     
pip              20.0.2    
pycparser        2.20      
PyInstaller      3.6       
PyQt5            5.13.0    
PyQt5-sip        4.19.19   
pyqt5-tools      5.13.0.1.5
python-dotenv    0.10.3    
pywin32-ctypes   0.2.0     
ui-to-py         1.2       
wheel            0.33.6    
whichcraft       0.6.1     
